backend/requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiohttp-retry==2.9.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asyncio==3.4.3
attrs==25.1.0
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distro==1.9.0
dnspython==2.7.0
fastapi==0.115.10
Flask==3.1.0
Flask-Sockets==0.2.1
frozenlist==1.5.0
gevent==24.11.1
gevent-websocket==0.10.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
MarkupSafe==3.0.2
multidict==6.1.0
openai==1.65.2
propcache==0.3.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pymongo==4.11.1
python-dotenv==1.0.1
requests==2.32.3
setuptools==75.8.2
sniffio==1.3.1
starlette==0.46.0
tqdm==4.67.1
twilio==9.4.6
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
websockets==15.0
Werkzeug==3.1.3
yarl==1.18.3
zope.event==5.0
zope.interface==7.2
